Update EC_CMD_CHARGE_CONTROL to version 2
Version 2 of EC_CMD_CHARGE_CONTROL can control battery sustainer.
It allows the host to set the upper and lower thresholds between
which the EC tries to keep the battery state of charge.

Version 2 of EC_CMD_CHARGE_CONTROL also supports 'GET' request.
It allows the host to query the current charge control settings.

localhost ~ # ectool chargecontrol
Charge mode = NORMAL (0)
Battery sustainer = off (-1% ~ -1%)

localhost ~ # ectool chargecontrol normal 66 66
Charge state machine is in normal mode with sustainer enabled.

localhost ~ # ectool chargecontrol
Charge mode = NORMAL (0)
Battery sustainer = on (66% ~ 66%)

localhost ~ # ectool chargecontrol normal
Charge state machine is in normal mode.

BUG=b:188457962
BRANCH=none
TEST=Atlas. See above.

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2929340

+++ include/charge_state_v2.h
@@ -52,8 +52,8 @@ struct charge_state_data {
 };
 
 struct sustain_soc {
-	int16_t lower;
-	int16_t upper;
+	int8_t lower;
+	int8_t upper;
 };
 
 /**
